construction of the local event channel handler object changed
The construction of the local event channel handler object was changed to a
form that needs fewer resources. Former it was a global static object that was
constructed via global initialization code. Now the construction is defered and
only the memory is reserved globally. The construction is done via the
placement new operator within the init function of the event channel handler
itself which is called from the main to initialize the famouso middleware.
Doing it that way avoids the need of the global initialization routines from the
compiler framework which leads to lesser resource utilization.



git-svn-id: https://svn-eos.cs.ovgu.de/repos/staff/mschulze/Research/MIKRO/famouso@856 235a714a-1e17-0410-b3bb-947d69a96898