add packed attribute to struct canID
The the description of the CAN id or the struct that describes the CAN id
respectively is mapped over a certain memory where a message with an id
resides. For holding the requirements and matching the correct value of the ID
it has to be packed as strong as possible in order to match with the value at
the underlying memory.


git-svn-id: https://svn-eos.cs.ovgu.de/repos/staff/mschulze/Research/MIKRO/famouso@1016 235a714a-1e17-0410-b3bb-947d69a96898