handling of endianness in subjects is continuous now
The subject of an event or event channel is always handled as string. If
you want to use a 64Bit UID you have to use and construct a temporary
famouso::mw::Subject this means that it is needed to call the
constructor explicitly whereby strings are handled automatically. The
endianness/byte order of a 64Bit value is handled automatically and the
64Bit value is stored in big endian format ever.

git-svn-id: https://svn-eos.cs.ovgu.de/repos/staff/mschulze/Research/MIKRO/famouso@1098 235a714a-1e17-0410-b3bb-947d69a96898