new etag binding protocol works collision free now
This is an enhancement of the etag binding protocol that works now
collision free. However, at the moment is not thread-safe, but in the
normal use-case we do not have threads. In future version, that can
change, hence, we need here further thinking to develop a version that
can handle such cases too. (fixed #13)

git-svn-id: https://svn-eos.cs.ovgu.de/repos/staff/mschulze/Research/MIKRO/famouso@1192 235a714a-1e17-0410-b3bb-947d69a96898