add generic command line parameter framework
The generic command line parameter framework contains a generator, that
can be used to generate a specific command line parameter option set for
a specific component. As an example, the UDPMultiCastNL is the first
one, that uses the new feature. It should be noted, that the command
line parameter infrastructure is only available if a component use this
feature. This means, it is an optional feature, and, thus, it is
ensured, that on the AVR no resources are wasted.

git-svn-id: https://svn-eos.cs.ovgu.de/repos/staff/mschulze/Research/MIKRO/famouso@1206 235a714a-1e17-0410-b3bb-947d69a96898