add explicit default constructor for CAN message object (MOB)
This patch solves a bug, that was arisen due to missing default
constructor, leading to wrong and not deterministic behaviour. The
default value of the ID was set sometime wrong, because its value was
dependent on the current stack layout. Now the constructor initialise
the ID with the value zero.

git-svn-id: https://svn-eos.cs.ovgu.de/repos/staff/mschulze/Research/MIKRO/famouso@1230 235a714a-1e17-0410-b3bb-947d69a96898