rewritten initialisation code allowing for automatic configuration/initialization
The initialization code is rewritten to allow automatic configuration of
the middleware. All tools have now a generic main implementation.  The
configuration description is evaluated in the init-functions of FAMOUSO.
In dependency of the content of the description the middleware is
initialized, respectively. For example, if the configuration contains a
typedef with a configured Gateway, the relevant actions are performed
for initializing a gateway component.

git-svn-id: https://svn-eos.cs.ovgu.de/repos/staff/mschulze/Research/MIKRO/famouso@1237 235a714a-1e17-0410-b3bb-947d69a96898