remove a boost include that caused compilation error for the AVRs
The removed include is in that case not needed (fixes #24). However, if
sometimes a component, also available for the AVR, needs boost includes,
this approach fails. An option to prevent this situation is using the
boost::compatibilty library for all platforms that are not provide
standard C++-headers. To get this complete done additional defines have
to be set at compile time.

git-svn-id: https://svn-eos.cs.ovgu.de/repos/staff/mschulze/Research/MIKRO/famouso@1241 235a714a-1e17-0410-b3bb-947d69a96898