changing the assignment of the return type of the process method
The return type of the process method is now determined by an additional
template parameter of the process method itself. Former, that was a
template parameter of the policy class. However, this had constrained
the use, because additional template parameters to policy class were not
possible. With, this change it is now possible, but the type needs to be
fully specified during the use within the IF_CONTAINS_TYPE_() generated
compile time selector classes.

git-svn-id: https://svn-eos.cs.ovgu.de/repos/staff/mschulze/Research/MIKRO/famouso@1254 235a714a-1e17-0410-b3bb-947d69a96898