PAXCAN -- a generic CAN framework
This patch adds the generic CAN framework PAXCAN. It specifies an
interface for accessing the CAN in always the same way. The framework is
used by FAMOUSO to abstract from the different low-level CAN driver
APIs. Thus, it is possible to write FAMOUSO independent of the concrete
underlying hardware. PAXCAN provides functionality for simulating
interrupts on devices, which are only support polling.

The intrgration of new drivers in the framework is easy, because they
have to fulfil only few requirements.

At the moment Peak CAN driver for Linux and Windows are supported.
SocketCAN, RTCAN and canary from the AVR follow.

git-svn-id: https://svn-eos.cs.ovgu.de/repos/staff/mschulze/Research/MIKRO/famouso@1280 235a714a-1e17-0410-b3bb-947d69a96898