etag binding protocol blocks the middleware core not any more
A new class is integrated, preventing the middleware core getting
stalled, in case of going to a blocking protocol. This is realized by
launching a new thread,  running the middleware core further. If the
blocking protocol is finished, the thread is joined.  Then former event
loop proceeds.

This new feature is needed on platforms only where asio is the core of
the event processing loop. Thus, it is necessary to enable the
additional treatment in such environments only. This activation is done
via a compile time selector trait that detects automatically in which
environment it runs.

This patch is related to issue #31 and fixes #31.

git-svn-id: https://svn-eos.cs.ovgu.de/repos/staff/mschulze/Research/MIKRO/famouso@1335 235a714a-1e17-0410-b3bb-947d69a96898