setting of callback of SubscriberEventChannel changed
The subscribe() method is put to the dummy callback only if there is no
other callback until now. Formerly, the callback was always set to a
dummy, because it was assumed that the callback was set after the
subscribe call by all users using FAMOUSO. However, this assumption is
false and can not be hold.

git-svn-id: https://svn-eos.cs.ovgu.de/repos/staff/mschulze/Research/MIKRO/famouso@1356 235a714a-1e17-0410-b3bb-947d69a96898