blocking local subscriber does not block publisher any more
One of the intended key features of FAMOUSO is having no control flow
dependencies between Publisher and Subscriber.

Until now it was a problem that local subscribers can block local
publishers if one subscriber does not return from its subscription
callback.  This was a contradiction to the key feature mentioned above.

While we cannot change this behaviour for AVR implementation, it is
possible to change it for the
EventLayerClientStub/EventLayerMiddlewareStub configuration and it is
provided with that patch.

git-svn-id: https://svn-eos.cs.ovgu.de/repos/staff/mschulze/Research/MIKRO/famouso@1364 235a714a-1e17-0410-b3bb-947d69a96898