add a configurable storage framework
The storage provides an iterator interface and the underlying storage
structure is configurable to arrays or dynamic/static lists.
Furthermore, it provides an uninitialized decorator class which allows
for deferring constructor calls and carry this job to the user, which
can use a placement new later for construction of objects.

git-svn-id: https://svn-eos.cs.ovgu.de/repos/staff/mschulze/Research/MIKRO/famouso@1469 235a714a-1e17-0410-b3bb-947d69a96898