add attribute filter framework
The framework allows defining of compile-time, and run-time filters.
Compile-time filters are very resource-efficient, but somehow difficult
to define for the most users. That is the point where run-time filters
come into play. Run-time filters can be defined in the same way as
compile-time filters, which is also difficult. However there is the
possibility to define filters via lambda-expression syntax which is very
handy and easy to understand. The patch adds, beside the framework
itself, also some test programs, where you can see the different forms
of defining a filter, and its usage too.

git-svn-id: https://svn-eos.cs.ovgu.de/repos/staff/mschulze/Research/MIKRO/famouso@1508 235a714a-1e17-0410-b3bb-947d69a96898