fix bug in famouso's idler component
In the Idler the mutex lock functions was called twice in the same
thread context. This is forbidden, because it could lead to deadlocks
thus the pthread lib throws an exception. Now we use a lock and a
condition variable. We wait on the condition that is given by a signal
handler if the program is aborted by the user.

git-svn-id: https://svn-eos.cs.ovgu.de/repos/staff/mschulze/Research/MIKRO/famouso@1535 235a714a-1e17-0410-b3bb-947d69a96898