Bugfix to EventLayerMiddlewareStub
The ECH uses async_write() on TCP stream sockets to publish
event to subscriber clients. Previously it was possible that
multiple async_write() requests got posted independently of
whether previous async_write()s where completed. In this case
the data stream got interleaved due to the fact that asio
does not support posting multiple async_write() in this way.
This patch adds an async_write() request queueing to fix this
problem. In addition it improves performance and reduces RAM
usage for a local publisher publishing to multiple local
subscribers.

git-svn-id: https://svn-eos.cs.ovgu.de/repos/staff/mschulze/Research/MIKRO/famouso@1613 235a714a-1e17-0410-b3bb-947d69a96898