Improves EventLayerMiddlewareStub subscriber crash detection/flow control
If data is published in a higher rate than the subscriber
can process it, the async write operation queue will flood
the memory. To avoid dropping data in case of high peak throughput
nothing is discarded before a deadline was passed. The deadline
is given relative to the last time no async write was pending.

git-svn-id: https://svn-eos.cs.ovgu.de/repos/staff/mschulze/Research/MIKRO/famouso@1616 235a714a-1e17-0410-b3bb-947d69a96898