AFP: Moves FEC code creation in FEC defragmenter
Calling fec_new() which calculates decoding tables for the
given code parameters is VERY EXPENSIVE for large fragment
counts. This commit moves the call from first arriving
fragment to the last needed one. Thus, it is only called
if there are enough fragments for reconstruction.
Additionally it does not block the subscriber callback for
the first callback of an FEC event which may result in
dropping successive fragments in the ECH.

git-svn-id: https://svn-eos.cs.ovgu.de/repos/staff/mschulze/Research/MIKRO/famouso@1617 235a714a-1e17-0410-b3bb-947d69a96898