fix broken gateway functionality introduced during AFP integration
In the r1606 patch set, the NetworkAdapter functionality changed in a
way that gateways could not distinguish between local and remote
subscriptions. The gateway assumed always local subscriptions and avoids
constructing forward channels for remote subscriptions. This patch
corrects the introduced wrong behaviour, thus gateways are now able to
determine the origin of subscroptions.

git-svn-id: https://svn-eos.cs.ovgu.de/repos/staff/mschulze/Research/MIKRO/famouso@1657 235a714a-1e17-0410-b3bb-947d69a96898