fixes wrong constructor calls in AttributeSet construction
The old scheme using template recursion for calling constructors has the
problem that the last constructor call uses an invalid this-pointer that
points one byte after the sequence's owned memory. I fixed the problem by
omitting constructor calls at all, using instead a static construct
method. Calling this method is always safe because it needs not a object
and thus no this-pointer is in the game, however the functionality of
constructing the sequence in the right manner is still given.
(fixes #77)

git-svn-id: https://svn-eos.cs.ovgu.de/repos/staff/mschulze/Research/MIKRO/famouso@1718 235a714a-1e17-0410-b3bb-947d69a96898