Adds possibility to pass special publish parameters to the ANL and below
NetworkGuard and/or AbstractNetworkLayer define a type PublishParamSet which is
passed to the layers above. A pointer to an instance of this type may be passed
to publish functions as an optional parameter.

Special publish parameters include real time stuff, like whether an event is
a real time event or the transmission time window, or QoS-Attributes needed
at lower layers, like the number of repetitions to mask omissions.

This feature is not yet supported for
EventLayerClientStub/EventLayerMiddlewareStub configurations.

git-svn-id: https://svn-eos.cs.ovgu.de/repos/staff/mschulze/Research/MIKRO/famouso@1944 235a714a-1e17-0410-b3bb-947d69a96898